requirements.txt
Flask
Werkzeug
Flask-SQLAlchemy
Flask-Login
SQLAlchemy
python-dotenv=